requirements.txt
future==0.18.3
customtkinter~=5.1.3
Pillow==10.0.1
keyboard~=0.13.5
gphoto2~=2.3.4
moviepy~=1.0.3
numpy==1.24.3
opencv-python~=4.10.0.82
CTkMessagebox~=2.0
py3dns~=3.2.1
git+https://gitea.ksol.io/karolyi/py3-validate-email@v1.0.9
